settings: setup.sh: don't fail with existing ~/bin/gvimless
git-svn-id: https://subversion.assembla.com/svn/magicant/settings@250 9a5e03f8-c55e-0410-a490-9d27d325e308

+++ setup.sh
@@ -85,6 +85,10 @@ if command -v vim >/dev/null 2>&1; then
 	vim/setup.sh
 	if command -v gvim >/dev/null 2>&1; then
 		#makelink vimless bin/gvimless
-		ln "$HOME/bin/vimless" "$HOME/bin/gvimless"
+		if [ -e "$HOME/bin/gvimless" ]; then
+			echo "~/bin/gvimless already exists"
+		else
+			ln "$HOME/bin/vimless" "$HOME/bin/gvimless"
+		fi
 	fi
 fi
